% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_decisions.R
\name{apply_futility_action}
\alias{apply_futility_action}
\title{Apply futility action to trial state}
\usage{
apply_futility_action(state, arm, action = "drop_arm")
}
\arguments{
\item{state}{Current trial state}

\item{arm}{Arm that hit futility}

\item{action}{Action to take: "stop_trial", "drop_arm", or "continue"}
}
\value{
Updated state and action taken
}
\description{
Apply futility action to trial state
}
