% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{check_conversion_trigger}
\alias{check_conversion_trigger}
\title{Check if conversion trigger is met}
\usage{
check_conversion_trigger(state, theta)
}
\arguments{
\item{state}{Current trial state}

\item{theta}{Hybrid design parameters}
}
\value{
Logical
}
\description{
Check if conversion trigger is met
}
