% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compile_hybrid_results}
\alias{compile_hybrid_results}
\title{Compile hybrid trial results}
\usage{
compile_hybrid_results(state, theta)
}
\arguments{
\item{state}{Final trial state}

\item{theta}{Hybrid design parameters}
}
\value{
List of trial metrics
}
\description{
Compile hybrid trial results
}
