% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ba_posterior_cpp}
\alias{compute_ba_posterior_cpp}
\title{Compute between-arm posterior probability (C++)}
\usage{
compute_ba_posterior_cpp(a_exp, b_exp, a_ref, b_ref)
}
\arguments{
\item{a_exp}{Posterior shape for experimental arm}

\item{b_exp}{Posterior rate for experimental arm}

\item{a_ref}{Posterior shape for reference arm}

\item{b_ref}{Posterior rate for reference arm}
}
\value{
P(HR < 1 | data)
}
\description{
Compute between-arm posterior probability (C++)
}
