% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ssr.R
\name{compute_ba_posterior_pwe_mc}
\alias{compute_ba_posterior_pwe_mc}
\title{Monte Carlo comparison for PWE posteriors}
\usage{
compute_ba_posterior_pwe_mc(
  a_exp,
  b_exp,
  a_ref,
  b_ref,
  interval_cutpoints,
  n_samples = 1000
)
}
\arguments{
\item{a_exp}{Posterior shapes for experimental arm}

\item{b_exp}{Posterior rates for experimental arm}

\item{a_ref}{Posterior shapes for reference arm}

\item{b_ref}{Posterior rates for reference arm}

\item{interval_cutpoints}{Interval boundaries}

\item{n_samples}{Number of Monte Carlo samples}
}
\value{
P(median_exp > median_ref)
}
\description{
Samples from posteriors and compares median survival times.
}
