% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_p_between_arm}
\alias{compute_p_between_arm}
\title{Compute between-arm posterior probability}
\usage{
compute_p_between_arm(state, theta, base_args)
}
\arguments{
\item{state}{Current trial state}

\item{theta}{Hybrid design parameters}

\item{base_args}{evolveTrial base configuration}
}
\value{
Posterior probability
}
\description{
P(HR_AB < 1 | data) where A is experimental, B is reference
}
