% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_p_single_arm}
\alias{compute_p_single_arm}
\title{Compute single-arm posterior probability}
\usage{
compute_p_single_arm(post_a, post_b, hist_hazard, hr_threshold, base_args)
}
\arguments{
\item{post_a}{Posterior shape parameters}

\item{post_b}{Posterior rate parameters}

\item{hist_hazard}{Historical hazard}

\item{hr_threshold}{Target HR threshold}

\item{base_args}{evolveTrial base configuration}
}
\value{
Posterior probability
}
\description{
P(HR < c | data) where HR = lambda_arm / lambda_hist
}
