% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_pp_curve}
\alias{compute_pp_curve}
\title{Compute predictive probability curve}
\usage{
compute_pp_curve(state, theta, base_args, scenario_params)
}
\arguments{
\item{state}{Current trial state}

\item{theta}{Hybrid design parameters}

\item{base_args}{evolveTrial base configuration}

\item{scenario_params}{Scenario parameters}
}
\value{
Data frame with n_add and pp columns
}
\description{
Compute predictive probability curve
}
