% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_pp_posterior}
\alias{compute_pp_posterior}
\title{Compute PP using posterior method (fast approximation)}
\usage{
compute_pp_posterior(state, n_add, theta, base_args)
}
\arguments{
\item{state}{Current trial state}

\item{n_add}{Additional patients per arm}

\item{theta}{Hybrid design parameters}

\item{base_args}{evolveTrial base configuration}
}
\value{
Predictive probability
}
\description{
Compute PP using posterior method (fast approximation)
}
