% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ssr.R
\name{compute_pp_posterior_single}
\alias{compute_pp_posterior_single}
\title{Compute posterior-based PP for single N_add}
\usage{
compute_pp_posterior_single(hr_hat, current_n, n_add, total_events, eff_ba)
}
\arguments{
\item{hr_hat}{Current HR point estimate}

\item{current_n}{Current total sample size}

\item{n_add}{Additional patients per arm}

\item{total_events}{Current total events}

\item{eff_ba}{BA efficacy threshold}
}
\value{
Conditional power (PP approximation)
}
\description{
Uses normal approximation to log(HR) to compute conditional power.
}
