% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_pp_predictive}
\alias{compute_pp_predictive}
\title{Compute PP using predictive method (Monte Carlo)}
\usage{
compute_pp_predictive(state, n_add, theta, base_args, scenario_params)
}
\arguments{
\item{state}{Current trial state}

\item{n_add}{Additional patients per arm}

\item{theta}{Hybrid design parameters}

\item{base_args}{evolveTrial base configuration}

\item{scenario_params}{Scenario parameters}
}
\value{
Predictive probability
}
\description{
Compute PP using predictive method (Monte Carlo)
}
