% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{compute_pwe_median}
\alias{compute_pwe_median}
\title{Compute median survival from PWE hazards}
\usage{
compute_pwe_median(lambda, interval_cutpoints)
}
\arguments{
\item{lambda}{Vector of hazard rates per interval}

\item{interval_cutpoints}{Interval boundaries}
}
\value{
Median survival time
}
\description{
Compute median survival from PWE hazards
}
