% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\docType{data}
\name{CONVERSION_TRIGGERS}
\alias{CONVERSION_TRIGGERS}
\title{Conversion triggers}
\format{
An object of class \code{character} of length 3.
}
\usage{
CONVERSION_TRIGGERS
}
\description{
Conversion triggers
}
\keyword{datasets}
