% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{count_arm_events}
\alias{count_arm_events}
\title{Count events for an arm up to analysis time}
\usage{
count_arm_events(state, arm, analysis_time)
}
\arguments{
\item{state}{Current trial state}

\item{arm}{Arm name}

\item{analysis_time}{Analysis time}
}
\value{
Integer event count
}
\description{
Count events for an arm up to analysis time
}
