% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{create_hybrid_state}
\alias{create_hybrid_state}
\title{Create initial hybrid trial state}
\usage{
create_hybrid_state(arm_names, reference_arm, theta, base_args)
}
\arguments{
\item{arm_names}{Character vector of arm names (e.g., c("Control", "Experimental"))}

\item{reference_arm}{Name of reference arm for between-arm comparison}

\item{theta}{Hybrid design parameters (see create_hybrid_theta)}

\item{base_args}{evolveTrial base configuration}
}
\value{
List containing trial state
}
\description{
Create initial hybrid trial state
}
