% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{draw_posterior_hazard_samples_r}
\alias{draw_posterior_hazard_samples_r}
\title{Draw posterior hazard samples (ORIGINAL R VERSION)}
\usage{
draw_posterior_hazard_samples_r(
  num_intervals,
  events_per_interval,
  person_time_per_interval,
  prior_alpha_params,
  prior_beta_params,
  num_samples = 1000
)
}
\description{
Draw posterior hazard samples (ORIGINAL R VERSION)
}
\keyword{internal}
