% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoint.R
\name{draw_posterior_response_rate}
\alias{draw_posterior_response_rate}
\title{Draw posterior samples for binary response rate}
\usage{
draw_posterior_response_rate(
  n_responses,
  n_total,
  alpha_prior = 1,
  beta_prior = 1,
  num_samples = 1000
)
}
\arguments{
\item{n_responses}{Number of responders}

\item{n_total}{Total number of patients}

\item{alpha_prior}{Beta prior shape1 parameter (default 1 for uniform)}

\item{beta_prior}{Beta prior shape2 parameter (default 1 for uniform)}

\item{num_samples}{Number of posterior samples to draw}
}
\value{
Numeric vector of posterior samples for response rate
}
\description{
Uses Beta-Binomial conjugacy to sample from the posterior distribution
of the response rate.
}
\details{
Prior: Beta(alpha_prior, beta_prior)
Likelihood: Binomial(n, p)
Posterior: Beta(alpha_prior + successes, beta_prior + failures)
}
