% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_analysis.R
\name{exp_arms_from_args}
\alias{exp_arms_from_args}
\title{Extract experimental arm names from an argument list}
\usage{
exp_arms_from_args(args)
}
\arguments{
\item{args}{Argument list containing \code{arm_names} and \code{reference_arm_name}.}
}
\value{
Character vector of experimental arm names.
}
\description{
Extract experimental arm names from an argument list
}
