% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\docType{data}
\name{FUTILITY_ACTIONS}
\alias{FUTILITY_ACTIONS}
\title{Futility actions}
\format{
An object of class \code{character} of length 3.
}
\usage{
FUTILITY_ACTIONS
}
\description{
Futility actions
}
\keyword{datasets}
