% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{get_historical_hazard}
\alias{get_historical_hazard}
\title{Get historical hazard for an arm}
\usage{
get_historical_hazard(arm, base_args, scenario_params)
}
\arguments{
\item{arm}{Arm name}

\item{base_args}{evolveTrial base configuration}

\item{scenario_params}{Scenario parameters}
}
\value{
Numeric hazard rate (or vector for PWE)
}
\description{
Get historical hazard for an arm
}
