% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{handle_sa_futility}
\alias{handle_sa_futility}
\title{Handle single-arm futility for an arm}
\usage{
handle_sa_futility(state, arm, theta)
}
\arguments{
\item{state}{Current trial state}

\item{arm}{Arm name that hit futility}

\item{theta}{Hybrid design parameters}
}
\value{
Updated state
}
\description{
Handle single-arm futility for an arm
}
