% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoint.R
\name{make_state_binary}
\alias{make_state_binary}
\title{Create state container for binary endpoint trial}
\usage{
make_state_binary(arm_names, max_total_patients_per_arm)
}
\arguments{
\item{arm_names}{Character vector of arm names}

\item{max_total_patients_per_arm}{Named integer vector of max N per arm}
}
\value{
State list for binary trial simulation
}
\description{
Create state container for binary endpoint trial
}
\keyword{internal}
