% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{ph_beta_mode_var}
\alias{ph_beta_mode_var}
\title{Compute mode and variance of log-HR posterior for PH model (DISPATCHER)}
\usage{
ph_beta_mode_var(
  E_C,
  PT_C,
  E_T,
  PT_T,
  alpha_prior,
  beta_prior,
  mu,
  sigma,
  tol = 1e-06,
  max_iter = 50
)
}
\description{
Compute mode and variance of log-HR posterior for PH model (DISPATCHER)
}
