% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{ph_beta_mode_var_r}
\alias{ph_beta_mode_var_r}
\title{Compute mode and variance of log-HR posterior (ORIGINAL R VERSION)}
\usage{
ph_beta_mode_var_r(
  E_C,
  PT_C,
  E_T,
  PT_T,
  alpha_prior,
  beta_prior,
  mu,
  sigma,
  tol = 1e-06,
  max_iter = 50
)
}
\description{
Compute mode and variance of log-HR posterior (ORIGINAL R VERSION)
}
\keyword{internal}
