% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_analysis.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot power versus type I error for calibration grids}
\usage{
plot_calibration(cal, target_alpha = 0.1, label_top_n = 3)
}
\arguments{
\item{cal}{List returned by \code{grid_calibrate()}.}

\item{target_alpha}{Type I error threshold to display.}

\item{label_top_n}{Number of feasible designs to annotate.}
}
\value{
A ggplot object.
}
\description{
Visualises the output of \code{grid_calibrate()} by plotting power against type I
error, optionally highlighting Pareto frontiers and feasible designs.
}
