% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoint.R
\name{prob_response_below}
\alias{prob_response_below}
\title{Calculate posterior probability that response rate is below threshold}
\usage{
prob_response_below(
  n_responses,
  n_total,
  threshold,
  alpha_prior = 1,
  beta_prior = 1
)
}
\arguments{
\item{n_responses}{Number of responders}

\item{n_total}{Total number of patients}

\item{threshold}{Response rate threshold}

\item{alpha_prior}{Beta prior shape1 parameter (default 1)}

\item{beta_prior}{Beta prior shape2 parameter (default 1)}
}
\value{
Posterior probability P(p < threshold)
}
\description{
P(p < threshold | data) using Beta posterior
}
