% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_driver.R
\name{release_cluster}
\alias{release_cluster}
\title{Release an evolveTrial simulation cluster}
\usage{
release_cluster(cluster)
}
\arguments{
\item{cluster}{A cluster object created by \code{create_simulation_cluster()} or
\code{parallel::makeCluster()}.}
}
\value{
NULL invisibly.
}
\description{
Stops and releases resources for a cluster created by
\code{create_simulation_cluster()}.
}
\seealso{
\code{\link[=create_simulation_cluster]{create_simulation_cluster()}}
}
