% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_hybrid_simulations_cpp}
\alias{run_hybrid_simulations_cpp}
\title{Run multiple hybrid trial simulations (C++ implementation)}
\usage{
run_hybrid_simulations_cpp(
  n_sim,
  theta_list,
  base_args_list,
  scenario_params_list
)
}
\arguments{
\item{n_sim}{Number of simulations}

\item{theta_list}{Design parameters}

\item{base_args_list}{Base configuration}

\item{scenario_params_list}{Scenario parameters}
}
\value{
Data frame with simulation results
}
\description{
Run multiple hybrid trial simulations (C++ implementation)
}
