% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{sample_vs_ref_medians_independent_r}
\alias{sample_vs_ref_medians_independent_r}
\title{Sample medians for vs-ref independent model (ORIGINAL R VERSION)}
\usage{
sample_vs_ref_medians_independent_r(slCtrl, slTrt, args, num_samples)
}
\description{
Sample medians for vs-ref independent model (ORIGINAL R VERSION)
}
\keyword{internal}
