% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{sample_vs_ref_medians_ph}
\alias{sample_vs_ref_medians_ph}
\title{Sample medians for vs-ref PH model (DISPATCHER)}
\usage{
sample_vs_ref_medians_ph(slCtrl, slTrt, args, num_samples)
}
\description{
Sample medians for vs-ref PH model (DISPATCHER)
}
