% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoint.R
\name{simon_final_efficacy}
\alias{simon_final_efficacy}
\title{Check Simon-style final efficacy rule}
\usage{
simon_final_efficacy(n_responses, r)
}
\arguments{
\item{n_responses}{Total number of responders}

\item{r}{Total response threshold (success if responses > r)}
}
\value{
TRUE if efficacious
}
\description{
In classical Simon design, declare efficacy if total responses > r
}
\keyword{internal}
