% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_endpoint.R
\name{simon_stage1_futility}
\alias{simon_stage1_futility}
\title{Check Simon-style stage 1 futility rule}
\usage{
simon_stage1_futility(n_responses, r1)
}
\arguments{
\item{n_responses}{Number of responders at stage 1}

\item{r1}{Stage 1 futility boundary (stop if responses <= r1)}
}
\value{
TRUE if should stop for futility
}
\description{
In classical Simon design, stop at stage 1 if responses <= r1
}
\keyword{internal}
