% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_hybrid_trial_cpp}
\alias{simulate_hybrid_trial_cpp}
\title{Simulate a single hybrid trial (C++ implementation)}
\usage{
simulate_hybrid_trial_cpp(theta_list, base_args_list, scenario_params_list)
}
\arguments{
\item{theta_list}{List of design parameters}

\item{base_args_list}{List of base trial configuration}

\item{scenario_params_list}{List of scenario parameters}
}
\value{
List with trial results
}
\description{
Simulate a single hybrid trial (C++ implementation)
}
