% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{simulate_pwe_survival}
\alias{simulate_pwe_survival}
\title{Simulate survival time from PWE model}
\usage{
simulate_pwe_survival(lambda, interval_cutpoints)
}
\arguments{
\item{lambda}{Hazard rates per interval}

\item{interval_cutpoints}{Interval boundaries}
}
\value{
Survival time
}
\description{
Simulate survival time from PWE model
}
