% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_pwe_survival_batch_cpp}
\alias{simulate_pwe_survival_batch_cpp}
\title{Simulate PWE survival times (C++ vectorized)}
\usage{
simulate_pwe_survival_batch_cpp(n, lambda, interval_cutpoints)
}
\arguments{
\item{n}{Number of survival times to simulate}

\item{lambda}{Hazard rates per interval}

\item{interval_cutpoints}{Interval boundaries (length = n_intervals + 1)}
}
\value{
Vector of n survival times
}
\description{
Simulate PWE survival times (C++ vectorized)
}
