% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_armadillo_random}
\alias{test_armadillo_random}
\title{Test Armadillo matrix operations}
\usage{
test_armadillo_random(n)
}
\arguments{
\item{n}{Matrix dimension}
}
\value{
Mean of random normal matrix
}
\description{
Test Armadillo matrix operations
}
