% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_rcpp_sum}
\alias{test_rcpp_sum}
\title{Test Rcpp setup}
\usage{
test_rcpp_sum(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Sum of the vector
}
\description{
Simple test function to verify Rcpp and RcppArmadillo are working.
}
