% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_trial.R
\name{update_posteriors}
\alias{update_posteriors}
\title{Update posterior parameters from trial data}
\usage{
update_posteriors(state, base_args)
}
\arguments{
\item{state}{Current trial state}

\item{base_args}{evolveTrial base configuration}
}
\value{
Updated state with posterior parameters
}
\description{
Update posterior parameters from trial data
}
