% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bayes_PE_v8_gemini__calendar_time.R
\name{calculate_interval_metrics_fast}
\alias{calculate_interval_metrics_fast}
\title{Recalculates events and person-time using a more efficient data.table approach.}
\usage{
calculate_interval_metrics_fast(patient_data, interval_cutpoints)
}
\description{
Recalculates events and person-time using a more efficient data.table approach.
}
