% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveTrial-package.R
\docType{package}
\name{evolveTrial-package}
\alias{evolveTrial}
\alias{evolveTrial-package}
\title{evolveTrial: Adaptive Clinical Trial Utilities}
\description{
Helpers for Bayesian adaptive platform/umbrella clinical trials, including simulation, decision rules, and reporting.
}
\author{
\strong{Maintainer}: Naim Rashid \email{naim@unc.edu}

}
\keyword{internal}
