% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_cpp_dispatchers.R
\name{sample_vs_ref_medians_independent}
\alias{sample_vs_ref_medians_independent}
\title{Sample medians for vs-ref independent model (DISPATCHER)}
\usage{
sample_vs_ref_medians_independent(
  slCtrl,
  slTrt,
  args,
  num_samples,
  ctrl_cache = NULL
)
}
\description{
Sample medians for vs-ref independent model (DISPATCHER)
}
