% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_ssr.R
\name{simulate_pwe_time}
\alias{simulate_pwe_time}
\title{Simulate survival time from PWE model}
\usage{
simulate_pwe_time(lambda, interval_cutpoints)
}
\arguments{
\item{lambda}{Hazard rates per interval}

\item{interval_cutpoints}{Interval boundaries}
}
\value{
Survival time
}
\description{
Uses inverse CDF method with piecewise constant hazard.
}
